added context menu tests to the unstable list
