+++ .travis.yml
@@ -7,17 +7,15 @@ env:
     - TERM=dumb
 
 script:
-    - "./gradlew clean -Dtest.single=ContextMenuTests test -i"
-    #- "./gradlew clean check -i"
+    - "./gradlew clean check -i"
 
 after_success:
     - "./gradlew jacocoRootReport coveralls"
 
 branches:
   only:
-    #- master
-    #- release
-    - 1256-mark-all-read
+    - master
+    - release
 
 before_install:
     - "export DISPLAY=:99.0"
